Add interface to control FRC mode
:Release Notes:
In order to reduce bandwitdth on VTG 4K input, FRC should
be turned off. So, add interface to control it.

:Detailed Notes:
HAL IG Update [yes]
Request to All SoC [yes]
Add interface to enable or disable FRC mode on vt driver
The default settings is FRC on mode.
To disable FRC, call this function with the parameter FALSE
before HAL_VT_Init/InitEx. This setting is reset when VSM
is connected to recovery on error.

:Issues Addressed:
[SCDCR-1231][HAL-API] New API to disable FRC
[KTASKWBS-7591][HAL-API] New API to disable FRC
[MSTARREQ-248]][HAL-API] New API to disable FRC

Change-Id: I2d1d065e4639f7b6bc21ac63ec8e5f68740fb7f2